Reorganized Tags/top-filter Warning now displayed if no result
