+++ src/style/pages/_index.scss
@@ -10,3 +10,7 @@ body {
   flex-wrap: wrap;
   gap: 5%;
 }
+
+.search-error_no-result {
+  font-size: larger;
+}
